refactor dot coordinates parsing
