+++ webknossos-datastore/app/com/scalableminds/webknossos/datastore/dataformats/zarr/ZarrDataLayers.scala
@@ -83,3 +83,17 @@ case class ZarrSegmentationLayer(
 object ZarrSegmentationLayer {
   implicit val jsonFormat: OFormat[ZarrSegmentationLayer] = Json.format[ZarrSegmentationLayer]
 }
+
+object ZarrCoordinatesParser {
+  def parseDotCoordinates(
+      cxyz: String,
+  ): Option[(Int, Int, Int, Int)] = {
+    val singleRx = "\\s*([0-9]+).([0-9]+).([0-9]+).([0-9]+)\\s*".r
+
+    cxyz match {
+      case singleRx(c, x, y, z) =>
+        Some(Integer.parseInt(c), Integer.parseInt(x), Integer.parseInt(y), Integer.parseInt(z))
+      case _ => None
+    }
+  }
+}
